First commit of TJL element onto master
This commit is after my master clean-up and upgrade to v140.

So to have a fresh solution, I started from Zero etjump repo and I add
the TJL function back.

Only one issue remain before the release.

So this fork is now up to date to the lates etjump 2.1.0 (all cpp)