fix the demo playback bug
Older demos were not working on a newer etjump version,
due to the introduced shifting in the entityType_t we
have made to add a new entity type in the game.
This patch reorderds entityType_t enum to allow older
demos to be played again, and also adds 2.3.0 specific
hack, to remap demo entity types to match the fixed
enum order.