Improvements to the comms between client & server
Implement a standardized serialization solution for commands from server
to client to ensure they're always in sync.
Refactored timerun view to only handle rendering and the timerun class
to handle the actual server commands parsing & state management.
Refactored the server to send overall record information (time,
checkpoint) instead of the relevant season. Relevant season is just
printed on server banner when a new record for the season is set