Feat/timerun v2 (#995)
* Add sqlite_modern_cpp dependency

Add the dependency and add a simple wrapper for handling migrations

* wip: implement logger & sync context v1

* fix log class

* Add support for adding seasons

* Add thread safe logging

* timerun v2 client connect & disconnect

* connect/disconnect/interrupt handling for timeruns

migrate db handling to a separate repository from the main timerun
handler

* Handle record completions

* first attempt at timerun timer based on aciz' code

* cleanup

* command parser & related utilities

* help text & edit season command

* add & edit season commands and /records command

* add support for importing records from old format

* fix compilation issues on linux

* enable manual ci runs

* Update ci.yml

* remove old timerun backend code

* cleanup

* cleanup

* record print improvements

* add positional args for command parser

improve records command to support both positional and non-positional
args

* fix overflow bug in command parser

* add loadcheckpoints command

* Ensure same run names with diff colors are not accepted

Technically run names are different if mapper uses ^1Run and ^2Run to
separate the names. However, this is difficult for the user and we also
strip the color codes when fetching the records so we cannot accept
them.

* simplify error prints for command parser

* cleanup

* improved prints for top record

* Proper records printing like it used to be

* Improvements to the comms between client & server

Implement a standardized serialization solution for commands from server
to client to ensure they're always in sync.
Refactored timerun view to only handle rendering and the timerun class
to handle the actual server commands parsing & state management.
Refactored the server to send overall record information (time,
checkpoint) instead of the relevant season. Relevant season is just
printed on server banner when a new record for the season is set

* Fix checkpoints renderer

* simplified checkpoint rendering

* remove unused code

* fix compiler issues & bugs

* include missing include

* add rankings per season & command to list seasons

---------

Co-authored-by: Zero <zero@etjump.com>