Optimize IRenderable system to reduce unnecessary computation (#1109)
IRenderable::beforeRender is now a boolean function, and IRenderable::render will not execute if it returns false. This makes it easier to optimize drawing functions by reducing the amount of complex math that needs to be done to setup renderable drawing in beforeRender, since we can just easily return false now to skip most of the stuff that's being done, and just skip the render that way.

This also fixes ob detector/watcher drawing while scoreboard is up, a small bug as a result of the recent refactor