Fix error messages for rtv [maplist] when no maps are available (#1240)
List might not necessarily be empty, but it's possible that the server doesn't have the maps installed or they are blocked via g_blockedMaps. Also clarify the error message about minimum available maps, as the current map in not necessarily a valid map if voting from a list which doesn't include the current map.