Fix projectiles getting stuck inside skyboxes
This is a vanilla bug/feature, where a grenade or other projectile can get stuck inside a skybox brush if the brush is thick and we happen to trace on a frame where we're fully inside it. This issue is exxagrated on high sv_fps values as we do more traces, so it's more likely that we get stuck. This makes projectiles ignore traces after they exit the playable area via skybox brush, so they are free to re-enter the playable area.