Add shortname options to the command system (#1338)
* Add shortname options to the command system

Each command parameter now has a name (e.g. --name) and a shortname (e.g. -n). --help can also be substituted with -h

Also reformats the command help prints slightly, the option description is now split to newline manually and indented if it goes over console maximum line length

* Fix tests

Dug up this wip branch from like 3 months ago and initially was planning on putting the shortname as the first param and adjusted the test for that, but then swapped it at some point it and forgot it seems.

* Fix tests 2 - electric boogaloo

* Where did this whitespace come from