Clear solid flag in G_InitGentity
gentity_t struct isn't wiped entirely for clients when they disconnect, and if a client reuses a slot which was previously on a client who was in a solid state, a client reusing that slot would become effectively solid while in spec initially, since spectators are never linked and their r.currentOrigin is never updated from the spectator spawnpoint. This causes other clients to think they are stuck inside a client on a spectator spawnpoint. Fortunately this doesn't really cause any harm in etjump with mostly nonsolid players.

Co-authored-by: ryzyk-krzysiek <ryzyk-krzysiek@users.noreply.github.com>