Refactor printer class
Renamed all printer functions to more concise names.

Added overloads to all printer functions, any function can not be called with an integer, gentity_t * or gclient_t * parameter. If the pointer is null, print is done to console.

Chat and popup messages printed to console are now automatically appended with a newline.