Utilities refactor + misc small bug fixes (#1433)
* Refactor printer class

Renamed all printer functions to more concise names.

Added overloads to all printer functions, any function can not be called with an integer, gentity_t * or gclient_t * parameter. If the pointer is null, print is done to console.

Chat and popup messages printed to console are now automatically appended with a newline.

* Deprecate old printing functions

Fixed several non-working admin system prints (quotes cannot be sent to chat via trap_SendServerCommand)

Fixed few incorrect printf format args

Fixed non-working prints for target_decay_use

* Remove useless CharPtrToString function

* Consolidate getPath from etj/g_utilities to etj_filesystem

* Deprecate G_MapExists

This was only used for checking if the map exists on the server regardless if it was playable -> use FileSystem::exists

Also fix FileSystem::exists leaving an open file handle - we don't have to register a file handle to check if a file exists.

* Cleanup

* Reorder banner switch + remove default label

* Add back default label

Since this is user-defined value, better to have handling for invalid arg

* Restore default behavior

Old behavior was to use regular banner as position for invalid arg

* Fix incorrect print destination on !listmaps

* Change wrapWords to take in a const reference

With this we can have Printer::console and Printer::log also take a const reference, and we no longer have to use std::move when calling either function with an lvalue

* Cleanup etj_motd a bit

* Remove redundant newlines from vote popups

* Oops was not supposed to push these

* Remove redundant newlines from goto/call/pm popups/chats

* Remove redundant newlines from specinvite/lock commands

* Remove redundant newlines from rtv popups