Add colorpicker for UI (#1422)
* Add colorpicker for UI

* Fix linux build

Not sure why this compiled for me but not in CI

* Tabs -> spaces for colorpicker menu

* Fix indentation in COLORPICKER_PREVIEW

* Remove some useless namespace scoping

These were initially inside _UI_Init so they were scoped to ETJump namespace, not required now that the initialization is extracted to a func in ETJump namespace

* Fix color picker capturing cursor outside the rect

If last focus item was color picker, any click on the screen would start capture again.

* Save/restore exact color cvar value on reset

If original cvar string was e.g. "white", pressing reset would rather set "1 1 1 1" value instead of "white". The color picker now restores the exact old cvar string when reset is pressed.

* Add toggle for normalized RGB sliders

Not the cleanest implementation, but hey it works (hopefully)