Display full maplist in map vote menu (#1431)
* Display full maplist in map vote menu

* Use const iterator

* Rename the cgame variable keeping track of request state

maplistRequested is more correct term as we are requesting it here, not receiving.

* Fix comment

* Only print arena file missing warnings in localhost

These are unnecessary warnings on an actual server, and the prints will flood the console if client doesn't have a lot of maps downloaded.

* Use case senstive comparison in Com_LocalPrintf

* Sort maplists alphabetically

* Simplify local file fetching, sort maps case insensitively

* Use const vector

* Newline

* Fix ui_restart breaking maplist

Re-send the request on UI load if we're already connected to a server