Separate savepos timer start from regular timer start
This way we can parse startTime as an int for savepos runs only, which is arguably a bit safer as it should never be negative in a real timerun. With this, it's only ever allowed to be negative for savepos runs, which in turns ensures that it's only negative when cheats are enabled, and records aren't saved.