Create a constructor for tokens
The current logger implementation was a bit dodgy, as it could potentially dereference null pointer for logger. It was initialized in loadTokens, which afaik is the only entry point to the token system, but it's better to be safe by initializing it in a constructor. Also make the token difficulty to string function part of the class so we can print out the difficulty with the admin commands for creating tokens.