Refactor JSON parsing + logging (#1463)
* Add JSON value parser

Safely parses JSON values and catches exceptions for invalid operations. Writes parsing errors to optional error variable. JSON key name can be provided for more descriptive error messages.

* Remove logger from JSON utils

It doesn't really make sense to have a logger for this, because the error would be printed as JSON-utils errors. Instead write errors to an optional error parameter, and have the function calling this be responsible for the error handling and printing.

* Use new value parser in chat replay

Also refactor the logging a bit

* Refactor MOTD class to use JsonUtils

* Refactor token JSON parsing + logging

Also make this whole system not throw G_Errors all the time, and just gracefully refuse to perform invalid actions

Also fixed !token move not positioning token to ground level

* Remove unused functions from etj_utilities.cpp

These are no longer needed after refactoring etj_tokens.cpp

* Refactor customvote JSON parsing & logging

Also fix up /listinfo message when no custom votes are defined on the server.

* Create a constructor for tokens

The current logger implementation was a bit dodgy, as it could potentially dereference null pointer for logger. It was initialized in loadTokens, which afaik is the only entry point to the token system, but it's better to be safe by initializing it in a constructor. Also make the token difficulty to string function part of the class so we can print out the difficulty with the admin commands for creating tokens.