Update customvotes for clients automatically during map (#1465)
* Update customvotes for clients automatically during map

If server edits custom votes using admin commands, or manually and executes 'readCustomvotes', invalidate current custom vote lists for all clients and have them request new ones the next time they open the vote menu.

Without this, the client would also crash if the customvote parsing failed on the server.

* Actually don't invalidate the cache

Not sure what I was thinking here originally.

* Close the menus if refresh happens while they're open

The data is garbage if the menus are open when the lists are refreshed.