Add bg print helper function to pmove
This is just a simple utility function for printing out values for both cgame and qagame for a given variable in Pmove. Often times when debugging Pmove related code, it's beneficial to print out values for both cgame and qagame Pmove, which this can be used for with a single function call.