Add savepos system (#1456)
* Add savepos system

Stores players position, angles, velocity, stance and timerun state into a file, which can then be restored using loadpos

Optionally takes a name and flags as arguments

Savepos is implemented fully in cgame side, meaning it can be used to save a valid state from demo playback (timerun state only works for demos recorded in ETJump 3.0.0 or newer)

Loadpos requires cheats to be enabled

* Remove redundant trap_AddCommand call

The arrays are automatically added as console commands

* Fix demo compatibility check

Semantically more correct this way

* Add readsavepos + improve loadsavepos messages

* Make logger class usable in cgame

Instead of writing to log file, prints out errors to console.

* Separate savepos timer start from regular timer start

This way we can parse startTime as an int for savepos runs only, which is arguably a bit safer as it should never be negative in a real timerun. With this, it's only ever allowed to be negative for savepos runs, which in turns ensures that it's only negative when cheats are enabled, and records aren't saved.

* Move serialize/deserialize functions to etj_savepos_shared.cpp

* Fix prone stance being saved as crouch

PMF_DUCKED is also set for prone, so prone needs to be checked first

* Add tests + exception handling to parsing

* Add missing return statement for out_of_range exception

* Fix comment typo + tweak tests

Check for too many args in a checkpoint argument too

* Fixup savepos command handler visibility

Teleport and timerun state restoration methods should be private

* Remove unused stripExtension function

Didn't end up needing this after all

* Add failsafe to JSON parsing

Don't crash the game if parsing fails.

* Remove cgame compatibility from Log class

Now that JSON utils doesn't do logging via this anymore, the cgame compatibility is not really needed.

* Ensure savepos deserialization never reads out of bounds

* Fix being unable to save a position that starts with a number

Forgot how Q_isNumeric works

* Respawn the player if they're dead while loading position

Similar to how /load works, respawn the player if they're dead while load happens to skip the death sequence

* Always interrupt ongoing timerun when using loadpos

If there's not timerun data associated with the savepos, then loadpos should reflect that state (no timerun active)

* Pluralize savepos parsing print

* Fix build

For development purposes, timerun state saving can be forcefully enabled with 'developer 1'

* Fix checkpoints

Checkpoints now carry the index of the checkpoint entity which we've hit, which makes it possible to know which checkpoint entities we've already activated during the run on client side.

* Flip timerun compatibility flag

Makes more sense to print it when it's not available

* Remove developer bypass from timerun state init

Was there for testing purposes