Add g_chatReplayMaxMessageAge (#1488)
* Add g_chatReplayMaxMessageAge

Can be set to > 0 to make chat replay messages expire. Any chat message older than current time - g_chatReplayMaxMessageAge <minutes< will be excluded from chat replay. The chat is still included in the replay file, it just won't be sent to clients.

* Make it CVAR_LATCH

Since chat messages are loaded to memory on game init, changing this without a restart can lead to unexpected behavior. For simplicitys sake, just make it a latch so we don't need to worry about re-parsing everything when it's changed.