Fix knife hitting nonsolid players + disable cursorhint
Knife has never been properly dealt with when it comes to nonsolid players, it has always hit them because the trace was never setup to handle that. This is no longer the case.

This also addresses a regression from the fireteam collision toggle implementation which caused knife backstab icon to draw for even nonsolid players.