Add experimental etj_smoothAngles for pmove-independent viewangles (#1495)
* Add experimental etj_smoothAngles for pmove-independent viewangles

This decouples refdef viewangles from ps.viewangles and instead calculates refdef angles from usercmd angles. This make it possible to update the client view more often than Pmove updates viewangles when pmove_fixed 1 is set. The tradeoff is that what you see is not exactly 1:1 to what physics are calculated with - for example at 250fps, every other viewangle update you see is effectively meaningless. This does not mess up with the real viewangles of the player, all physics calculations are still done with the proper values.

This is disabled if the player is prone, mounted or using a set weapon, from what I understood and tested, these do not work correctly since they require accurate pmext values, which are only updated when Pmove is run. Without disabling this, refdef angles would start to gradually get out of sync with the real ps.viewangles. It might be possible to do somehow, but it's not really critical so I'll come back to it some day if I feel like it. Teleportation events (teleports, load, portals, respawns) force refdef values back to playerstate values, to sync up the changes, as refdef doesn't know about these otherwise.

This is disabled by default, because I'm not 100% confident that this is entirely bug free, hence I'm treating it as experimental for now.

* Remove unnecessary oldViewAngles variable

This was needed for earlier revision where I tried to fix prone/mounted viewangles

* Add missing syncs to intermission & wounded state

We need to sync to playerstate with these too, rather than just exiting.

* Fix typo in comment

* Fix incorrect vector func for cg.xyspeed calculation