Fix bullet impact particle effects working unreliably (#1503)
* Fix bullet impact particle effects working unreliably

This was caused by a div by 0 fix done earlier in #1355, but that overwrote valid trace results which caused impact particles to not generate properly. To be honest I am still not fully sure what is actually wrong here, but we now just skip the particle generation if the second trace didn't hit anything, when shooting at water.

* Cleanup