Fix autoload/deathcam skip load spawn position
Autoload and respawn performed during the death sequence would initially spawn the player at a normal spawnpoint, run ClientThink and only teleport player to the saved position at the end of the frame. This caused the player to for example touch any triggers that are directly on top of spawnpoint, before getting placed into the correct position, which could cause entities to fire unexpectedly when joining a team or forcing a respawn during death sequence.