Fix target_init + timeruns not resetting player ammo correctly (#1515)
When players weapons were removed by target_init or timerun start, the ammo for the weapons wasn't cleared. This caused an issue with weapons that do no use clips, where picking up the weapon after it was removed would restore the ammo count the player had when the weapon was taken away, instead of having whatever ammo the picked up weapon had.