Add changelog to ETJump menus (#1532)
* Setup changelog parsing for CMake

Doing this in CMake means we don't need to actually bundle the changelog.md with the mod

* Add changelog menu item

ETJump -> Changelog displays the changelog from a given version (selectable via dropdown menu) in a scrollable textbox.

This also fixes a bunch of hardcoded values from the dropdown menu implementation, making it much more portable outside of the ETJump settings menu layout.

* Remove unused CHANGELOG_HEADERS macro, add a header guard to version_headers.h

The macro is not needed, initially thought I'd need the headers too but that was not the case.

Including a header guard to version_headers.h since it gets included in ui_main.cpp for reading the changelog files.

* Fix some inconsistencies with changelog formatting

Make some of the older entries have proper indentation like newer ones.

Handle bolded text before transforming line starts from * to -, to not break lines starting with ''* **X**'

* Use ifdef header guards

UI parsers don't support #pragma once

* Handle carets in changelog

Replaces them with a color escape so they are displayed normally

* Remove ETJump prefix from the dropdown menu entries

Just having the numeric version is cleaner and enough.

* Slightly reduce the max line width for changelog

* Remove unused COMBOLEFT macro

* Re-enable CI runs on changelog.md changes

Now that it's used to parse the changelog assets, it makes sense to run CI whenever it's updated.

* Tweak changelog background colors

Use slightly brighter background color in FUI, the readability is fine without the darker background in menus.

* Add changelog formatting documentation

* Fix ci.yml for PRs

Was supposed to remove changelog.md instead of .clang_format from the ignored list