Properly fix changemodel script action
The script action worked fine initially, but the intended use case was to change the models of brush entities using 'model2' key. This is stored in 'ent->s.modelindex2', as opposed to 'ent->s.modelindex'. Perform the swap in 'ent->s.modelindex' for 'misc_gamemodels', but use 'ent->s.modelindex2' for other entities.