Fix etj_autoSprint not working < 125FPS with pmove_fixed 1
Because 'pmove_fixed 1' forces 8ms command times (with 'pmove_msec 8'), running at < 125FPS results in the same usercmd being used for multiple calls to Pmove. Because we flipped the bit inside PmoveSingle, this would result in the bit getting flipped potentially multiple times within a single physics frame, which caused the state to not reflect the users actual inputs. Do the bit flipping before we call Pmove, on both client and server, in order to not flip the bit multiple times for a single usercmd.