Fix timerun entity validation
This was never functional because ent->runName wasn't set for start/stoptimers, and would fail most of the time anyway as the container search was done with a case sensitive string comparison, and radiant creates entity names with all lowercase letters. This also moves the prints from log to server console, as these warnings are mostly meant for mappers, so it makes sense to print them to console instead of log to make them easier to catch.