Automatic demo playback via demoqueue
Allows users to specify a directory and automatically play back all demos from that directory. This breaks with 'vid_restart' and 'ui_restart', but because there's no syscalls/other communication methods with engine to detect when these commands are executed, we can't do anything about it. There's also no way to detect a demo ending, so we just have to hook the autoplay to shutdown calls.