Revert "Fix autoload/deathcam skip load spawn position (#1516)"
This reverts commit da0cfac12271a6a10c886942348177857df47159.

This breaks things too much and doesn't make much sense when you really think about it. Players are supposed to spawn on spawn points, not in arbitrary locations in the map. The purpose of this cvar was never to change that, but rather just execute 'load' command automatically when you join a team. Having autoload bypass the whole process of spawning into a spawnpoint makes it impossible for mappers to create maps which assume that certain set of actions are performed for the player once they spawn (e.g. trigger a target_init entity). There's no perfect solution to that atm either, but I am working on something in regard to that.