Fix ent->s.onFireEnd timestamp on respawn (#1637)
Flamethrower timestamps are not cleared when the player respawns (even in vanilla ET), and 'P_WorldEffects' keeps applying flamethrower damage if the player was burning right before they died, and respawned within the flamethrower burn timer (2s). Spawnshield normally hides this in VET and negates the damage, but since we don't have spawnshield, the burn damage would carry over to respawn and keep damaging the player. Set the burn end timestamp to be in the past to stop the damage from carrying over (similar to what 'G_BurnTarget' does for entities which are not supposed to be set on fire).