Fix chatwindow enter key not working on 2.60b
The call to 'trap_Cmd_ExecuteText' on UI keyhandling to reset the chat window state to cgame was missing a newline at the end, which broke the commandline buffer on 2.60b and caused chat messages to not be sent properly. Also add missing newline to custom vote request and writeconfig, though these didn't seemingly cause problems given their execution context.