Add tracker_not_eq_any and tracker_not_eq_all keys for trackers (#1614)
This deprecates 'tracker_not_eq' key on tracker entities and replaces it with two new keys: 'tracker_not_eq_any' and 'tracker_not_eq_all'. This provides mappers with easy way to implement a 'tracker_not_eq' check with both AND + OR logic, where the clients tracker values must not match any or all of the specified values.

Tracker mapscripting doesn't support multiple indices in a single command, so there's no need to handle any of this there. There is simply an 'abort_if_equal' command which acts as an equivalent of 'not_eq' functionality. Mapscripting provides a lot more tools for conditional execution compared to standard entities, so implementing checks for multiple indices/values there is not worth the added complexity.