Add target_ft_setrules entity + supplementary ws keys (#1625)
* Add target_ft_setrules entity + supplementary ws keys

Allows setting fireteam rules for the activcators fireteam. Bypasses 'noftnoghost' and new 'noftsavelimit' & 'noftteamjumpmode' worldspawn keys, so mappers have precise control over when these rules are set. Spawnflag 1 makes it so that only the leader of the fireteam is able to activate the entity.

* Change savelimit & leader_only_message to pointers and alloc the strings

This way we can use custom string names and have \'set\' script action still work properly.

* Fix comment