Add etj_logCenterPrint for center print logging (#1627)
When enabled, prints centerprints to console, excluding save messages.

Consecutive, identical messages are not logged as long as the message re-appears while the previous message is still up (as determined by 'cg_centertime'). Once the message has faded from the screen, identical message may be logged again.

This bypasses 'cg_centertime 0', messages are always logged if the cvar is enabled, even if they are not shown on screen.