Add target_spawn_relay for executing actions on respawns (#1636)
This should be the be-all and end-all solution for firing entities/executing mapscript whenever a client respawns in a map.

The current tools for doing this are lackluster and all come with some sort of compromise. You have two options:

* Cover the spawnpoint(s) in a trigger that fires the targets/script events you want
* Target the spawnpoint(s) to entities you want to activate

Both of these methods are flawed and come with compromises:

* Covering spawn point with a trigger means you must perform this for all spawn points in the map.
* Targeting entities with a spawnpoint triggers entities twice when joining a team from spectators, due to way respawn is handled. You must also target all spawn points in the map to the entities you want to fire.

The solution to this mess is this entity, 'target_spawn_relay'. This is a "passive" entity that just sits in the map, and gets activated every time a player respawns in the map. Unlike spawnpoint targeted entities, this correctly only fires once when joining a team from spectators. This does not fire on revives, only fresh spawns.

The main usage is via these keys:

* 'team' - comma-separated list of teams that this relay fires for. For example, 'team axis,allies' would activate whenever players respawn on either axis or allied team. If this key is not set, the entity fires for all teams, including spectators. The valid team names are "axis", "allies" and "spectator".
* 'scriptname' - name of the scriptblock to activate. This calls the 'activate' trigger, and also supports 'activate axis/allies'. Activator is passed to the mapscript, so client-only actions may be performed. You are not required to set this, but it's an option if you want to perform mapscript actions.
* 'target' - targetname of entities which you want to activate while spawning.

You can only have a single relay per team in a map. If you have a relay which does not specify a team, that is your one and only relay that you can use. If you try to spawn additional relays with same team defined in the 'team' key, or multiple relays with no team set, the game will throw an error.