Refactor pmove utils to a global object
The main motivations for this are to reduce the amount of boilerplate required to setup fake pmove for renderables, and to improve performance. Rather than having every single renderable that requires a fake pmove for drawing to call a static method which runs 'PmoveSingle', there's now a global object that has a getters which return pointer to 'pmove_t' or 'usercmd_t' structs, owned by the pmove utils class. The actual fake pmove is ran after 'CG_PredictPlayerstate', but before 'CG_DrawActive' is called, so it's always valid and up-to-date before renderables use it. It's only ran if there's a renderable enabled that requires the fake pmove.

Performance wise, this is roughly a 5% improvement to fps on spec/demo playback, based on my test scenarios. Nothing massive, but consistent and measurable.

There some more simplifications that could still be done to some of the renderables with this new setup, but this is good enough for now.