Make static dlights fully client sided (#1668)
* Make static dlights fully client sided

Any dlight that doesn't have a targetname, scriptname, spawnflags or sound can be fully processed on client side. This allows more than 16 dlights in a map, if the dlights don't require any server-side processing. There's no real limit to these (arbitrarily set to 1024), but the renderer has a limit of 32 dlights in a scene, and will drop any extra dlights. This should be documented in the entity definition file.

* Add demo compatibility layer

* Support sound on clientside dlights

We just have to send the entity origin and 'entityNum -1' to 'trap_S_StartSound' in order to play the sound from the correct location. Because these are fully static, we don't need the entity number to track position and state to potentially cut-off/move the sound origin.