Fix chat replay potentially writing a null JSON object
If the replay file is empty/doesn't exists, and no chats were written during a session, a null JSON object was written to the file, which then failed to parse on load. This wasn't causing any issue per se, as the next valid message to store would just overwrite the null object, but better not write that to begin with.