Rework client authentication
Preliminary work for improved client authentication. Can be turned on with the CMake option 'NEW_AUTH'. This isn't "wired up" on server side yet, there is only basic message parsing done on GUID request response, and as such, using this will break all server-side authentication for clients. The main thing that this does for now is to prepare the client side for a more robust authentication system, specifically when it comes to hardware IDs - there is more than one hash sent to the server for identification (for Windows and Linux, macOS is... macOS, I can't test it).