Add manual GUID migration
Clients can issue 'migrateGuid' command to manually request the server to migrate their old GUID to a new one. This is useful in scenarios where a client has connected to a server once without their old GUID file present, but wants to migrate later. 'migrateGuid -f/--force' can be used to force the server to perform the migration again, even if the user database already contains an old GUID for the user, if they for example accidentally migrate using a wrong GUID.