Initial user database migration to new schema
Basic steps for migrating old user database to a new database. This doesn't do anything yet, old database is simply migrated to new database on init. The schema is very likely not final, but I'll figure that out once I start writing the code that actually interacts with the database, this is just an estimate on how I want to organize everything.