Fix sanitize() handling color strings incorrectly (#1687)
Explicitly check for a string being a color string rather than just greedily removing every '^' + following character, to avoid strings with escaped carets from being incorrectly sanitized. This might have some unexpected breakage as this is quite widely used in the code base, but luckily this is not too common of a scenario, and it has worked fine for the most part.