Don't try to perform authentication on demo playback
This is actually not handled in the old system either, though the likelihood of demo containing auth requests is very low, it would only really happen if a client had instant demo recording set on something like 'autoexec_default.cfg'.

This should probably be directly added to the 'ClientCommandsHandler' class at some point, so commands could be subscribed to such that the responses are omitted on demo playback.