IP ban checks implemented
Checks for both IPv4 and IPv6. Since we now track the last known IP address for clients, a client cannot evade IP ban by simply switching from IPv4 to IPv6 or vice versa.

Was contemplating whether I should load these to memory on init or not, but in the end I decided to do database access, as this is only checked on first connect, and keeping bans in memory would require careful cache management.