Fix an edge case with GUID migration failing to add a new user
If a user had never connected to the server (no old or new GUID in the database), authentication would fail as new GUID was not found in the database, but GUID migration would also fail as the old GUID was not found in the database either. Add a new user in such scenario.