Make portal teleports predicted
Portal teleports are now fully client side predicted. This works even if the destination is not in the current PVS, by abusing 'SVF_PORTAL' flag, similar to 'trigger_teleport_client'. The downside is that this brings back the bug where you might see an area through a skybox that is completely out of your PVS. I might add a cvar to disable prediction in the future (can be done on client side), if this is too bothersome for people.