Move portal debugging to client side
Deprecates 'g_portalDebug' server cvar in favor of 'etj_portalDebug' client cvar. Since we're calculating the bbox for portals on client anyway, there's no point in creating a temp entity on server and sending it over network for client to draw. This also fixes portal bboxes not drawing on flat floors, because the origin was inside a solid.

The cvar isn't in the menus, and I'm not quite sure if I am gonna add it, since it's a cheat cvar and I don't think there's currently a good way to make menu cvars not do anything if cheats are disabled, but I'll look into it some other time. This is a debugging cvar anyway so it's not so critical.

This also cleans up the railtrail function signatures and removes the unnecessary clientInfo param.