Add .clang-tidy file
This is more of a guideline file rather than a strict ruleset, following this to a tee is not expected or required. For example, this flags all uses of 'const_cast' and 'reinterpret_cast', and while they are fine to use (and a necessity for interoperability with old code without massive refactoring), the purpose really is to make you think whether you should be doing something differently to avoid doing it. It can be a bit loud, especially with old etmain code, but it's reasonable with newer code. This will probably evolve a bit in the future as I only skimmed through roughly 10 random files to determine the checks used.