Add etj_useExecQuiet for map/team autoexecs (#1695)
* Add etj_useExecQuiet for map/team autoexecs

When enabled, map and team-specific autoexecs are executed using 'execq', to omit the print about executing a config file from the console output.

Hidden from the menus on 2.60b as it does not support 'execq' (which is why we can't just enable this implicitly)

* Change the cvar to be a bitflag

This lets users select individually if they want to use silent exec for map and/or team autoexecs, and leaves an easy way to expand this in the future with other autoexec configs, if added.