Add cvar to disable menu flashing
'etj_noMenuFlashing' can be used to disable the lightning effect from the menu background shader. The lightning effect is especially problematic for people with photosensitive epilepsy, and it's just in general quite annoying.

This also adds a new category to 'General' settings tab called 'UI', which is where this setting is located. 'etj_menuSensitivity' was also moved here from underneath 'Client' submenu.