Add quick connect window to main menu
Adds a window to the right side of the main menu, which can hold up to five "quick connect" servers. The data is stored in 'quickconnect.dat' file in etjump directory. The interface allows for adding, editing, deleting and connecting to servers, and displays the server name (with ability to set a custom name), current map and player count. The list is not automatically refreshed, and requires a manual refresh to fetch up-to-date data. The menu can be turned off with 'etj_drawQuickConnectMenu' cvar.